Code refactoring to SwapChainManager
