+++ src/ComputeSharp.UI/ComputeSharp.UI.projitems
@@ -11,6 +11,7 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Controls\ComputeShaderPanel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\ComputeShaderPanel.Properties.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Helpers\SwapChainManager.Rendering.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helpers\SwapChainManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Helpers\DynamicResolutionManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)IShaderRunner.cs" />
